Moves & renames
Move code that's generally useful into the auth namespace
