+++ src/hellotest/HelloApp.java
@@ -10,9 +10,9 @@ import org.slf4j.LoggerFactory;
 
 import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.Logger;
-import us.kbase.auth2.service.mustache.MustacheProcessor;
-import us.kbase.hello.LoggingFilter;
-import us.kbase.hello.UncaughtExceptionHandler;
+import us.kbase.auth2.service.LoggingFilter;
+import us.kbase.auth2.service.template.mustache.MustacheProcessor;
+import us.kbase.auth2.service.ExceptionHandler;
 
 //TODO WAIT accept json in text/plain and application/x-www-form-urlencoded or manually handle it
 
@@ -26,7 +26,7 @@ public class HelloApp extends ResourceConfig {
 		register(MustacheMvcFeature.class);
 		property(MustacheMvcFeature.TEMPLATE_BASE_PATH, "templates");
 		register(LoggingFilter.class);
-		register(UncaughtExceptionHandler.class);
+		register(ExceptionHandler.class);
 		register(new AbstractBinder() {
 			@Override
 			protected void configure() {
